nginx: correctly handle 0-length data writes
When proxying a request, if the response uses chunk encoding but is
0-length, we send a 0-length DATA frame at the end to close the stream.

However if the stream is blocked and the headers can't be sent, the
0-length write is attempted, but is then accidentally considered as
complete because the NGX_AGAIN code returned in this case is
interepreted the same way as a 0-length write.

This makes NGINX explicitly check for NGX_AGAIN, and loop to the next
buffer in the chain only after it is determined that the stream is not
blocked.

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>