nginx: generate HTTP/3 response headers once
Whenever a request is prevented from sending response headers (e.g. due
to flow control), the sending is deferred at a later time. However every
time sending is attempted, a new array of response headers (to be passed
to quiche) is generated from scratch.

This is inefficient and could potentially lead to large amounts of
memory being allocated as the headers are allocated on the request's
memory pool.

Instead only generate the headers list once, and re-use it across
attempts.

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>