nginx: improve HTTP/3 request body handling
This updates the NGINX patch for the edge-triggered Data event.

Before this patch the HTTP/3 request body handling worked in a similar
way to HTTP/2's. Notably whenever a Data event was received, the stream
data would be read through quiche and buffered internally, until an
upstream event would read it. This didn't work very well because the way
the HTTP/3 implementation works doesn't actually match the HTTP/2 one in
this regard.

In HTTP/2 a single stream of multiplexed data is read from a single
socket. Meaning that different frames across different streams could be
read into the same input buffer.

In HTTP/3 however, different streams are abstracted in a similar way to
different socket (except that instead of using the socket API, the
quiche API is used).

This makes the HTTP/3 implementation more similar to that of HTTP/1.*,
where instead of data being pushed from the protocol implementation, the
data is actually _pulled_ (say, from the upstream module) when needed
(e.g. when the upstream connection becomes writable).

As far as HTTP/3 is concerned, the main complication comes from the fact
that the HTTP/1.* request body handlingd assumes that we know the size
of the body chunk being read beforehand (either becuase a content-length
header is provided, or due to chunk encoding's chunk size prefixes).
Through the quiche API, if not content-length header is provided, we
don't know how much data we are going to get until we get a Finished
event.

But when NGINX is reading a request's body (via c->recv()), it assumes
that if less data than what is expected is returned, the request is
blocked, while it could simply be that the stream is finished. This
means that to actually process the Finished event correctly, we need to
explicitly trigger a read event on the main connection such that it is
polled for HTTP/3 events again.

In addition, when a Finished event is received, we also need to trigger
a body filter run manually (as there may be data that was previously
read but not yet processed), as well as running the affected request
again. Both of these are done through posted events.

Incidentally this also makes request buffering just work out-of-the-box.

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>