h3: rename max_header_list to max_field_section
The name of this setting was changed in draft -28 but we never went and
updated it. Since the setting is exposed externally in the API, it could
confuse people that come to implement the more recent drafts. This
change fixes things.

Note that this does not update the qlog crate, since that spec is
lagging behind too; see: https://github.com/quicwg/qlog/issues/175

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>